+++ app/src/expo/expo.rs
@@ -1,9 +1,11 @@
 use crate::{
     error::CustomError,
-    get_push_notification_receipts::{
-        get_push_notification_receipts, ExpoPushReceipt, ExpoPushReceiptId,
+    get_push_notification_receipts::get_push_notification_receipts,
+    object::{
+        expo_push_message::ExpoPushMessage, expo_push_receipt::ExpoPushReceipt,
+        expo_push_receipt_id::ExpoPushReceiptId, expo_push_ticket::ExpoPushTicket,
     },
-    send_push_notifications::{send_push_notifications, ExpoPushMessage, ExpoPushTicket},
+    send_push_notifications::send_push_notifications,
 };
 
 pub struct Expo {
