+++ app/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod error;
 mod expo;
 mod get;
+mod object;
 mod post;
 pub use expo::expo::Expo;
 pub use get::get_push_notification_receipts;
