+++ app/src/object/expo_push_error_ticket.rs
@@ -1,0 +1,9 @@
+use serde::{Deserialize, Serialize};
+
+use super::details::Details;
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct ExpoPushErrorTicket {
+    pub status: String,
+    pub message: String,
+    pub details: Option<Details>,
+}
