+++ app/src/object/expo_push_success_recept.rs
@@ -1,0 +1,6 @@
+use serde::Deserialize;
+
+#[derive(Debug, Deserialize, PartialEq, Clone)]
+pub struct ExpoPushSuccessReceipt {
+    pub status: String,
+}
