+++ app/src/object/expo_push_success_ticket.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+pub struct ExpoPushSuccessTicket {
+    pub status: String,
+    pub id: String,
+}
